pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ire-rag-llm"
version = "1.1.0"
description     = "Local FAISS + RAG demo"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "python-dotenv>=1.1.0",
  "langgraph>=0.6",
  "langchain>=0.3.27",
  "langchain-openai>=0.1.7",
  "langchain-google-genai>=2",
  "langchain-community>=0.3",
  "langchain-huggingface>=0.3",
  "langchain-text-splitters>=0.3",
  "langchain-core>=0.3.72",
  "beautifulsoup4>=4.12.2",
  "huggingface_hub>=0.34",
  "sentence-transformers>=2.2.2",
  "faiss-cpu>=1.11.0",
  "torch>=2.2",
]

#tells setuptools to look inside /src
[tool.setuptools.packages.find]
where = ["src"] 

#creates the two console-scripts
[project.scripts]
run-retrieval = "rag.run_retrieval:main"
run-rag       = "rag.run_rag:main"
